Fix type mismatch in handle_syn_event
Cast evt.time.tv_sec to i64 to match last_activity type.
Add #[allow(clippy::unnecessary_cast)] since tv_sec is i64 on x86_64
but i32 on ARM (Kobo target), making the cast necessary for cross-platform compatibility.

+++ crates/core/src/input.rs
@@ -538,7 +538,8 @@ fn handle_syn_event(
     proto: TouchProto,
     ty: &Sender<DeviceEvent>,
 ) {
-    let tv_sec = evt.time.tv_sec;
+    #[allow(clippy::unnecessary_cast)]
+    let tv_sec = evt.time.tv_sec as i64;
     if (tv_sec - *last_activity).abs() >= 60 {
         *last_activity = tv_sec;
         ty.send(DeviceEvent::UserActivity).ok();
